functions/link2_poll/requirements.unexpanded.txt
azure-storage-file-share==12.3.0
dataclasses==0.6
google-cloud-secret-manager==2.1.0
google-cloud-storage==1.35.0